+++ opencog/embodiment/Control/OperationalAvatarController/PsiModulatorUpdaterAgent.cc
@@ -169,7 +169,7 @@ void PsiModulatorUpdaterAgent::publishUpdatedValue(Plaza & plaza,
 
     // Pack time stamp and all the modulator values in json format 
     Object jsonObj; // json_spirit::Object is of type std::vector< Pair >
-    jsonObj.push_back( Pair("timestamp", timeStamp) );
+    jsonObj.push_back( Pair("timestamp", (uint64_t) timeStamp) );
 
     foreach (Modulator & modulator, this->modulatorList) {
         jsonObj.push_back( Pair( modulator.getModulatorName(), modulator.getModulatorLevel() ) );
